Time: 12 ms (12.90%), Space: 9.7 MB (23.83%) - LeetHub

+++ 35-search-insert-position/35-search-insert-position.cpp
@@ -1,21 +1,19 @@
 class Solution {
 public:
     int searchInsert(vector<int>& nums, int target) {
-        int low=0;
-        int high=nums.size()-1;
+        long long low=0;
+        long long high=nums.size()-1;
         while(low<=high){
-        int mid=(low+high)/2;
+            int mid=(low+high)/2;
             if(nums[mid]==target){
                 return mid;
-                
             }
             else if(nums[mid]<target){
-                low=mid+1;
-            }
+low=mid+1;}
             else{
                 high=mid-1;
             }
-            }
+        }
         return low;
     }
 };
